+++ spiffworkflow-backend/src/spiffworkflow_backend/api.yml
@@ -622,6 +622,39 @@ paths:
             application/json:
               schema:
                 $ref: "#/components/schemas/ProcessModel"
+                
+  /process-model-import/{modified_process_group_id}:
+    parameters:
+      - name: modified_process_group_id
+        in: path
+        required: true
+        description: modified id of an existing process group
+        schema:
+          type: string
+    post:
+      operationId: spiffworkflow_backend.routes.process_models_controller.process_model_import
+      summary: Imports a process model from uploaded files.
+      tags:
+        - Process Models
+      requestBody:
+        required: true
+        content:
+          multipart/form-data:
+            schema:
+              type: object
+              properties:
+                files:
+                  type: array
+                  items:
+                    type: string
+                    format: binary
+      responses:
+        "201":
+          description: Process model imported successfully.
+          content:
+            application/json:
+              schema:
+                $ref: "#/components/schemas/ProcessModel"
 
   /process-model-tests/run/{modified_process_model_identifier}:
     parameters:
