+++ spiffworkflow-frontend/test/browser/test_process_model_import.py
@@ -55,21 +55,14 @@ def test_process_model_import(browser_context: BrowserContext):
     expect(url_input).to_be_visible()
     url_input.fill("https://github.com/sartography/example-process-models/tree/main/examples/0-1-minimal-example")
     
-    # Wait for validation to complete (look for the checkmark or preview data)
-    # We might need to wait a bit for the backend to validate the URL
-    page.wait_for_timeout(2000)  # 2 second timeout
-    
-    # No custom ID input - backend will automatically handle ID generation
-    
     # Click the import button
-    import_model_button = page.get_by_test_id("import-button")
-    expect(import_model_button).to_be_enabled()
-    import_model_button.click()
+    import_button = page.get_by_test_id("import-button")
+    expect(import_button).to_be_enabled()
+    import_button.click()
     
     # Wait for the import to complete and redirect to the process model page
     # ID will be automatically generated by the backend
     # The imported model will likely have a name based on the directory name in the URL
-    # For the example repository, this would be "0-1-minimal-example" or "0-1-minimal-example-{timestamp}"
     page.wait_for_url(re.compile(r"process-models/import-test-group/.+"))
     
     # Verify we're on a process model page
@@ -91,13 +84,6 @@ def test_process_model_import(browser_context: BrowserContext):
     # Clean up the imported process model
     # Go back to process models list
     page.goto(f"http://localhost:7001/process-groups/import-test-group/process-models")
-    
-    # Find the imported process model and delete it
-    # This part depends on the UI for deleting process models
-    # We're assuming there's a delete button or menu option
-    
-    # Example cleanup using command line (will be run after the test)
-    # The actual test will rely on the ./bin/clean_up_sample_process_models script
     
     # Verify we're back at the process models list
     expect(page).to_have_url(re.compile(r"/process-groups/import-test-group/process-models$"))
